+++ src/Repository/Model/Query.php
@@ -24,6 +24,7 @@ declare(strict_types=0);
 
 namespace Ampache\Repository\Model;
 
+use Ampache\Config\ConfigurationKeyEnum;
 use Ampache\Module\Authorization\Access;
 use Ampache\Config\AmpConfig;
 use Ampache\Module\System\AmpError;
@@ -1255,7 +1256,7 @@ class Query
         // filter albums when you have grouped disks!
         if ($this->get_type() == 'album' && !$this->_state['custom'] && AmpConfig::get('album_group')) {
             $album_artist = (array_key_exists('album_artist', $this->_state['sort'])) ? " `artist`.`name`," : '';
-            $final_sql .= " GROUP BY" . $album_artist . " `album`.`prefix`, `album`.`name`, `album`.`album_artist`, `album`.`release_type`, `album`.`mbid`, `album`.`year` ";
+            $final_sql .= " GROUP BY" . $album_artist . " `album`.`prefix`, `album`.`name`, `album`.`album_artist`, `album`.`release_type`, `album`.`mbid`, `album`.`year`, `album`.`original_year` ";
         } elseif (($this->get_type() == 'artist' || $this->get_type() == 'album') && !$this->_state['custom']) {
             $final_sql .= " GROUP BY `" . $this->get_type() . "`.`name`, `" . $this->get_type() . "`.`id` ";
         }
@@ -1915,7 +1916,11 @@ class Query
                         $this->set_join('left', '`artist`', '`song`.`artist`', '`artist`.`id`', 100);
                         break;
                     case 'year':
-                        $sql = "`album`.`year`";
+                        if (AmpConfig::get(ConfigurationKeyEnum::ALBUM_USE_ORIGINAL_YEAR)) {
+                            $sql = "`album`.`original_year`";
+                        } else {
+                            $sql = "`album`.`year`";
+                        }
                         break;
                 } // end switch
                 break;